requirements.txt
torch
pyyaml
torchtext
portalocker>=2.0.0
spacy
torchsummary